* Initial auto-build scripts
The post-update hook must be installed with a working copy.
Then the run_queue.sh script should be crontabbed to be ran
every 5 minutes. It has a lock around it, so it will not
spawn more than 1 build process at a time.

Tags pushed to the working copy will be queued to be built
after 10 minutes. This gives time for the tag to be updated
and prevents wasting too much resources.

Source is stored in the SRC_PATH and binaries are stored in BIN_PATH.
The source is built from 'make distrib' which currently only
works properly on v1.2.16-rc1 and up.